Added readonly capability for ADMIN, MOD and USER
Removed unused code

+++ src/main/java/com/atom/application/services/UserRoleService.java
@@ -1,5 +1,7 @@
 package com.atom.application.services;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -15,6 +17,8 @@ import org.springframework.stereotype.Service;
 @Service
 public class UserRoleService {
 
+    private final static String READONLY_ROLES[] = { "USER", "MODERATOR", "ADMIN" };
+
     @Autowired
     private UserRoleRepository repo;
 
@@ -24,6 +28,15 @@ public class UserRoleService {
 
     public UserRole addNewRole(UserRole role) {
         return repo.save(role);
+    }
+
+    public UserRole getUserRoleByName(String name) {
+        Optional<UserRole> persistedOpt = repo.findByName(name);
+        if (!persistedOpt.isPresent()) {
+            throw new EntityNotFoundException("No user role found with name: '" + name + "'");
+        }
+        UserRole persisted = persistedOpt.get();
+        return persisted;
     }
 
     public List<UserRole> getAllUserRolesByNames(List<String> requestedRoleNames) {
@@ -39,10 +52,22 @@ public class UserRoleService {
     }
 
     public void removeRoles(List<UserRole> roles) {
+        List<String> roleNames = roles.stream().map(UserRole::getName).collect(Collectors.toList());
+        List<String> readonlyNames = new ArrayList<>(Arrays.asList(READONLY_ROLES));
+        readonlyNames.removeAll(roleNames);
+        if (!readonlyNames.isEmpty()) {
+            throw new IllegalArgumentException("Cannot remove readonly user roles");
+        }
         repo.deleteInBatch(roles);
     }
 
     public UserRole updateRole(String existingRoleName, UserRole role) {
+        List<String> readonlyNames = Arrays.asList(READONLY_ROLES);
+        if (readonlyNames.contains(existingRoleName)) {
+            throw new IllegalArgumentException(
+                    "Cannot modify readonly user role with name: '" + existingRoleName + "'");
+        }
+
         Optional<UserRole> persistedOpt = repo.findByName(existingRoleName);
         if (!persistedOpt.isPresent()) {
             throw new EntityNotFoundException("A user role with name '" + existingRoleName + "' does not exist");
